backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
passlib
bcrypt==3.2.2
pydantic
pytest
pytest-cov
httpx